+++ lib/openstudio-analysis.rb
@@ -1,15 +1,17 @@
+# Ruby libraries to include
 require 'json'
+require 'securerandom'
+require 'logger'
+require 'pathname'
+
+# gems to always include
 require 'faraday'
-require 'uuid'
 require 'roo'
 require 'rubyXL'
 require 'erb'
 require 'zip'
 require 'semantic'
 require 'semantic/core_ext'
-require 'logger'
-require 'pp'
-require 'pathname'
 
 # core
 require 'openstudio/analysis/server_api'